# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Asset Manager API
  version: 0.1.0
  description: |
    API multi-proyecto de benchmarking y validación de evidencias.
    PR1 (V2 base): auth, proyectos, miembros, invitaciones, catálogo de KPIs,
    evidencias CRUD, jobs (placeholder) y API keys.
servers:
  - url: /api
    description: Mismo origen (Vite proxy en dev, Nginx en prod)
...
paths:
...
  /projects/{projectId}/evidencias:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - in: query
        name: kpi_id
        schema: { type: string, format: uuid }
      - in: query
        name: empresa_comparable
        schema: { type: string }
      - in: query
        name: decision_final
        schema: { type: string }
      - in: query
        name: search
        schema: { type: string }
    get:
      tags: [evidencias]
      operationId: listEvidencias
      responses:
        "200":
          description: Lista
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Evidencia"
    post:
      tags: [evidencias]
      operationId: createEvidencia
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEvidenciaBody"
      responses:
        "201":
          description: Creada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Evidencia"
  /projects/{projectId}/evidencias/{evId}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - in: path
        name: evId
        required: true
        schema: { type: integer }
    get:
      tags: [evidencias]
      operationId: getEvidencia
      responses:
        "200":
          description: Detalle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Evidencia"
    patch:
      tags: [evidencias]
      operationId: updateEvidencia
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEvidenciaBody"
      responses:
        "200":
          description: Actualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Evidencia"
    delete:
      tags: [evidencias]
      operationId: deleteEvidencia
      responses:
        "204":
          description: Eliminada
...
  /projects/{projectId}/evidencias/import:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    post:
      tags: [evidencias]
      operationId: importEvidencias
      summary: Subir un XLSX y calcular el diff (status="previewed")
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                mode:
                  type: string
                  enum: [upsert, replace]
                  default: upsert
      responses:
        "201":
          description: Preview generado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvidenciaImportRunPreview"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationFailed"
  /projects/{projectId}/evidencias/imports:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    get:
      tags: [evidencias]
      operationId: listEvidenciaImports
      responses:
        "200":
          description: Historial de imports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EvidenciaImportRun"
  /projects/{projectId}/evidencias/imports/{runId}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - in: path
        name: runId
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [evidencias]
      operationId: getEvidenciaImport
      responses:
        "200":
          description: Detalle (incluye diff completo)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvidenciaImportRunWithDiff"
    delete:
      tags: [evidencias]
      operationId: discardEvidenciaImport
      responses:
        "204":
          description: Descartado
  /projects/{projectId}/evidencias/imports/{runId}/commit:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - in: path
        name: runId
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [evidencias]
      operationId: commitEvidenciaImport
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                dry_run: { type: boolean, default: false }
                confirm_project_name:
                  type: string
                  description: Obligatorio si mode=replace.
      responses:
        "200":
          description: Resultado del commit
          content:
            application/json:
              schema:
                type: object
                required: [summary, applied]
                properties:
                  summary:
                    $ref: "#/components/schemas/EvidenciaImportSummary"
                  applied: { type: boolean }
  /projects/{projectId}/evidencias/template.xlsx:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    get:
      tags: [evidencias]
      operationId: downloadEvidenciasTemplate
      responses:
        "200":
          description: Plantilla XLSX
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema: { type: string, format: binary }
  /projects/{projectId}/evidencias/download.xlsx:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    get:
      tags: [evidencias]
      operationId: downloadEvidenciasXlsx
      responses:
        "200":
          description: Export simétrico
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema: { type: string, format: binary }
...
components:
  parameters:
    ProjectId:
      in: path
      name: projectId
      required: true
      schema:
        type: string
        format: uuid
  responses:
...
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    ValidationFailed:
      description: Validation Failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  schemas:
...
    ApiError:
      type: object
      required: [error, code]
      properties:
        error: { type: string }
        code:
          type: string
          enum:
            - UNAUTHENTICATED
            - FORBIDDEN
            - INSUFFICIENT_ROLE
            - PROJECT_NOT_FOUND
            - PROJECT_SLUG_TAKEN
            - KPI_NOT_FOUND
            - KPI_DUPLICATE
            - EXCEL_PARSE_FAILED
            - EXCEL_NO_CHANGES
            - INGESTION_NOT_FOUND
            - EVIDENCE_NOT_FOUND
            - INVITATION_INVALID
            - INVITATION_EXPIRED
            - VALIDATION_FAILED
            - RAG_UNAVAILABLE
            - AGENT_FAILED
            - INTERNAL_ERROR
        details:
          type: object
          nullable: true
        traceId: { type: string }
...
    Evidencia:
      type: object
      required:
        [id, project_id, empresa_comparable, fuente_tipo, created_at]
      properties:
        id: { type: integer }
        project_id: { type: string, format: uuid }
        kpi_id: { type: string, format: uuid, nullable: true }
        empresa_comparable: { type: string }
        entidad_fuente: { type: string, nullable: true }
        ano: { type: integer, nullable: true }
        indicador: { type: string, nullable: true }
        categoria_efqm: { type: string, nullable: true }
        fuente_nivel: { type: string, nullable: true }
        fuente_tipo: { type: string }
        fuente_titulo: { type: string, nullable: true }
        url_validada: { type: string, nullable: true }
        ubicacion_fuente: { type: string, nullable: true }
        texto_evidencia: { type: string, nullable: true }
        valor_reportado: { type: number, nullable: true }
        unidad: { type: string, nullable: true }
        comparabilidad: { type: string, nullable: true }
        observacion_metodologica: { type: string, nullable: true }
        decision_final: { type: string, nullable: true }
        definicion_referencia: { type: string, nullable: true }
        unidad_base_referencia: { type: string, nullable: true }
        indicador_fuente: { type: string, nullable: true }
        encaje_indicador: { type: string, nullable: true }
        estado_auditoria: { type: string, nullable: true }
        tipo_compania: { type: string, nullable: true }
        unidad_estandarizada: { type: string, nullable: true }
        valor_estandarizado: { type: number, nullable: true }
        created_at: { type: string, format: date-time }
    CreateEvidenciaBody:
      type: object
      required: [empresa_comparable, fuente_tipo]
      properties:
        kpi_id: { type: string, format: uuid, nullable: true }
        empresa_comparable: { type: string }
        entidad_fuente: { type: string, nullable: true }
        ano: { type: integer, nullable: true }
        fuente_nivel: { type: string, nullable: true }
        fuente_tipo: { type: string }
        fuente_titulo: { type: string, nullable: true }
        url_validada: { type: string, nullable: true }
        ubicacion_fuente: { type: string, nullable: true }
        texto_evidencia: { type: string, nullable: true }
        valor_reportado: { type: number, nullable: true }
        unidad: { type: string, nullable: true }
        comparabilidad: { type: string, nullable: true }
        observacion_metodologica: { type: string, nullable: true }
        decision_final: { type: string, nullable: true }
    UpdateEvidenciaBody:
      allOf:
        - $ref: "#/components/schemas/CreateEvidenciaBody"
...
    EvidenciaImportSummary:
      type: object
      required: [filename, file_hash, mode, totals]
      properties:
        filename: { type: string }
        file_hash: { type: string }
        mode:
          type: string
          enum: [upsert, replace]
        totals:
          type: object
          required:
            - rows_in_file
            - parse_errors
            - new
            - updated
            - unchanged
            - kpi_not_found
            - will_remove_in_replace
          properties:
            rows_in_file: { type: integer }
            parse_errors: { type: integer }
            new: { type: integer }
            updated: { type: integer }
            unchanged: { type: integer }
            kpi_not_found: { type: integer }
            will_remove_in_replace: { type: integer }
    EvidenciaImportRun:
      type: object
      required: [id, project_id, status, mode, filename, created_at, summary]
      properties:
        id: { type: string, format: uuid }
        project_id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        filename: { type: string }
        file_hash: { type: string }
        mode:
          type: string
          enum: [upsert, replace]
        status:
          type: string
          enum: [previewed, committed, discarded, failed]
        summary:
          $ref: "#/components/schemas/EvidenciaImportSummary"
        error: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        committed_at: { type: string, format: date-time, nullable: true }
    EvidenciaImportRunWithDiff:
      allOf:
        - $ref: "#/components/schemas/EvidenciaImportRun"
        - type: object
          required: [diff]
          properties:
            diff:
              type: object
              additionalProperties: true
    EvidenciaImportRunPreview:
      type: object
      required: [run, summary]
      properties:
        run:
          $ref: "#/components/schemas/EvidenciaImportRun"
        summary:
          $ref: "#/components/schemas/EvidenciaImportSummary"
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: assetmgr.sid
security:
  - sessionCookie: []